Improve Landscape Support
* Add QR Screen Landscape Support

Update the QR tabs so that they support
landscape mode.

* Tweak FlowTerminalPage

Update the FlowTerminalPage so that
the content scrolls more consistently.

Also added missing SafeArea Widgets to the
issuance and sign flow and the history
overview screen.

* Improve Landscape Support

Add SafeArea widgets throughout the app,
mainly to improve landscape support on iOS
throughout the app. As the bottom 'gesture'
bar would otherwise overlap with a bunch
of content.