Navigate based on has_registration
* Refactor CheckIsAppInitializedUseCase

Rename from CheckIsAppInitializedUseCase
to IsWalletInitializedUseCase for consistency.

* Update UseCase Name

Add explicit UseCase suffix

* Update to relative import

* Introduce WalletErrorHandler

When no onError callbacks are set
the app does not crash when an error
from the wallet_core is (re)thrown.

This commit introduces a very simple
error handler class which causes the app
to crash when uncaught exceptions are
detected. Giving the developer instant
feedback of any unhandled/unexpected
mistakes. #failfast

* Clean up api.rs

- Split variable assignment and
return statement in the has_registration
api
- Simplify return type of create_wallet

* Clean up TypedWalletCore

Previously added a Completer to track
the initialization process, however
with the recent tweaks to the wallet_core
this does not seem to be needed. So removing
for sake of simplicity.

* Update Splash Navigation

Update the SplashBloc so that it provides
info on the initialization and availability
of the PID, so that the SplashScreen can pick
the correct destination for the user.

* Consistent Naming + Docs

Align the flutter code with the rust api
by renaming walletCreated (and similar) calls
to reflect the has_registration() api.

Also added some documentation to the repository
and usecase classes to make it a bit more explicit
what is actually (expected) to happen.

* Await Wallet Initialization

The first thing the app does (on the splash)
is checking if the wallet is already registered.
This can happen too soon, meaning the call is made
before the rust core is initialized. Calls to the
rust core are now shielded by a complete.future to
make sure the walletCore has always finished initializing
before interacting with it.

* Replace isInitializedStream with isWalletCreated()

Align what the WalletRepository exposes a bit more
with what is coming from the RustCore.

* Expose has_registration() api

Introduce the has_registration api in the
flutter rust api bridge, which can be used
to check if a wallet account has been created
(i.e. if a pin has been successfully configured)