repair  introduction screen tests
* change after review

* changed and removed some introduction screen tests

* Perform cleanup

- Fix formatting
- Remove unused code
- Update comment

* Update Introduction Goldens

- Fix issue with conflicting scroll
controllers by providing a default inside
the introduction pages
- Fix issue with resolving the next button
for the golden tests
- Update the goldens to match the new design

* Update locale

* Update Privacy Footer

* Move BottomSection

Render the bottomSection in the
introduction pages by means of a
footer parameter, so that we know
where the optional scroll needs to
end and the paging animation looks
cleaner.

* Update Introduction

- Introduce IntroductionPrivacyPage
- Pull progressStepper into the
IntroductionPage

* Add Subtitle Support

Add subtitle support to the introduction
page so that it's rendered in the onboarding.

* Update Onboarding Images

* Update introduction titles & add subtitles

* Remove 1 introduction page

* Remove app disclaimer intro page