Implement tracing output for Android
For printing log statements on Android, we use just the low-level access
provided by the "android_logger" statement. Alternatively, we could use
the logging implementation that this crate provides (which provides
output for using the "log" crate) and configure "tracing" to output
using that (using this crate's "log" feature), but then we lose
consistent formatting between iOS and Android.