Update Watermark
* Add kWhite0 Color

Add a white with 0 opacity
in favor or the transparent
color as the latter causes
a slight blackish border which
is unwanted.

* Clean up CardHolograph

- Add some documentation
- Fix the startPosition of the
holograph overlay so that it
defaults to a nice position when
the phone is held in the user's hand
- Wrap foils in Roll to provide some
defaults

* Add Introduction Skip Animation

* Update Watermark Colors

Results of a 1-on-1 session with
Stijn to live code and test changes
to the Foil setup and adjust the
assets.