Update Verification/Issuance/Sign Transition
* Export Constant

* Update Issuance/Sign Screen Animations

Make sure the initial transition is also skipped
for the issuance and sign flows.

* Clean up imports

* Tweak FakePagingAnimatedSwitcher

Disable the duration by tweaking the animation
times. This makes sure that:
- The old child is immediately hidden.
- When navigating from a state where the
in animation was disabled, the old child is
still neatly animated out of the screen.

* Remove Initial Delays

Remove the mock delays, smoothening
out the opening animation of the
sign/verify/issue flows.
Also slightly tweaked the animation to
not perform the paging on the initial
load and fixed a crash caused by conflicting
access to the primary scroll controller when
aborting the flow from the first screen.

* Update Transition Method Name

Make it slightly more explicit
by aligning it with the enum name.

* Add Custom Transitions

Update the SecuredPageRoute to support
custom transitions.

* Fix 2 UI Errors

- Dedicated scroll controller conflicted
with the PrimaryScrollController, fixed
by removing it.
- Added missing margin below the close button
on the FlowTerminalPage