Prepare wallet_app for DigiD OIDC Flow
* Propagate process_uri Errors

Instead of having a dedicated 'unknown'
enum type when a uri can't be processed
by the rust_core, we now simply return
an Error which neatly ends up in the
stream.

* Implement Feedback

Implement more PR Feedback in the
wallet_app codebase.

* Remove stale generated files

* Implement Rust MR Feedback

* Update Mock Names

Make it consistent with the previous
repo rename.

* Implement MR Feedback

* Refactor DigiD to Digid

* Improve DeeplinkService & Test

Make methods async so they can be awaited
and thereby properly tested.

* Start testing DeeplinkService

* Introduce Mock Factory

Easily resolve mocks using Mocks.create()

* Clean up core_wallet_repository_test

* Analysis fix

* Mock with Mockito and UseCase Tests

- Centralize mocks used in tests in the
wallet_mocks.dart file, and built mocks
using Mockito (vs. manual creations).
- Add tests for the auth usecases

* Improve CoreWalletRepository test #2

* Improve CoreWalletRepository test

* Extract TypedWalletCore Interface

Extract the interface to make classes that
rely on it more testable.

* Test AppLifecycleService

* Run Cargo Fmt

* Introduce DigiDAuthenticationRepository Test

* Reset wallet_constant

* PR Tweaks

- Revert tweak test account_server_url and pubkey
- Revert fake_paging temp fix
- Initialize TypedWalletCore at provider

* Move UriFlowEventExtension

Ideally this is generated when the serde generator
is ran, but I can't seem to get that to work, so making
the extension part of the wallet_app codebase.

* Migrate to Bincode

Use bincode (de)serialize to transfer
uri flow events from rust to flutter.

* Add Docs & Usecases

- Add some documentation to the deeplink_service.dart
- Avoid referencing Repositories directly, instead
route the calls through dedicated Usecases.

* Improve Uri Process Stream

Make the stream logic a bit more readable.

* Delay Uri Processing

Delay the Uri processing until the app
is in the resumed state. This makes sure
that the Uri is not processed too soon in
some cases, as not adding this check could
trigger the situation where:
- The app is in the background
- The lock timer is expired, meaning the
app will now be locked
- A deeplink triggers the app to open
- The deeplink is instantly processed
because the `locked` boolean is flipped
on app open, and not while it's in the
background (UNWANTED!)
- By debouncing on the applifecyclestate
first, this situation is avoided, as it
makes sure the locked flag is always
in the correct state before using it
to debounce the uri processing.

* Delay ProcessUri Calls

Delay processing any deeplink Uris
until the app is unlocked, this makes
sure don't process any uri's before
we can actually handle the result.

Might need an extension at some point
to allow deeplinks that we do want to
process before unlucking the app, but
so far I can't think of any relevant
usecase for that so sticking to this
(imo) simple and clean solution.

* Clean up Duplicate Enum

The AuthStatus and AuthenticationStatus
enums really represented the same thing,
so opted to remove one for now.

* Add localization

Localize hardcoded strings in the
personalization flow.

* Consistent DigiD naming

Make sure we consistently use
'DigiD' with a capital trailing
D everywhere (in stead of 'Digid').

* Update GetDigidAuthUrlUseCaseImpl

Make it use the AuthRepo in favor
of calling the core directly.

* Use register API

- Use the real register api (requires
a running wallet_provider)
- Unlock the wallet with any pin
(useful while testing, added a TODO)
- Do not automatically revert to idle
in DigiDstate so that the event
can't be missed while the user is
entering her pin

* Extract authStatusUpdate handler

* Improve AuthState Update Propagation

Make sure the UI updates and is in the
correct state when the app is cold
started with a authentication deeplink.

* Basic 'process_uri' flow setup

Setup the wallet_app so that it passes
on the uri to the wallet_core which
now returns a stream. This stream is
observed and relevant events are passed
on to the related repository so that the
wallet_app can continue the relevant flow.

* Introduce CallbackStream

Setup a generic, static stream in
the wallet_core that can be used
to communicate back to flutter at
any given time. Currently used to
notify that a placeholder deeplink
has been processed and the digid
login flow can continue.

* Update Android launchMode

Update the launchMode so that (e.g.
when a deeplink is clicked) the existing
activity is always brought to the front
instead of creating a new one. This is
desired because otherwise a new Flutter engine
might be created, which also triggers a
duplicate initialization of the WalletCore,
which in term causes the app to crash.

* Pass on Deeplink Uri

Delegate deeplink uris that can't be
handled by the wallet_app directly
to the wallet_core so they can be
handled there.

* Introduce GetDigiDAuthUrl

Get the digid auth url through the
wallet_core and pass it on to the
WalletPersonalizeScreen so it can be
launched.