Move 'lock' flag to wallet_core
* Regenerate bridge

Regenerate the bridge because the order of the
methods has changed, causing the ci to break.

* Group Wallet & WalletLockSink

* Re-order api.rs

Keep the is_valid_pin api at the top,
since it is the only api (right now)
that doesn't interact with the wallet.

* Add instrument annotation

* Rename locked property

Make sure naming is consistent between
the property and the getter.

* Fix clippy warning

* Update wallet_app tests

* Change lock behavior

Locking should never fail, as this would
likely imply some broken security. As such
crash the app when it does.

* Run fmt & Add docs

* Add lock mechanism test

Verify that the lock()/unlock()
results in updated wallet.lock state

* Fix Deeplink Issue

Since the lock state is now coming out of the
wallet_core, there is also a slight delay since
it's coming over an async bridge. This delay is
just enough to trigger an issue in the deeplink
processing logic, causing the processUri method
to trigger before the device is unlocked (or actually
it triggers before it realizes it was ever locked).

To circumvent this, for now I added a fixed debounce
delay before checking if the app is locked. Would
be nice to avoid this delay in the future, but for
now opting this solution as sufficient since worst
case the deeplink is dropped or queued for later
processing. So no crucial (security) implications.

* Simply WALLET_LOCK_SINK reference

Instead of keeping it in a RwLock just
keep a direct reference to the StreamSink
as I currently don't oversee any concurrency
issues.

* Stream lock state from wallet_core

Make the init() method on the wallet_core
return a Stream that fires when the wallet
lock state changes.

* Implement placeholder (un)lock

* Define initial core (un)lock api