Call hardware keystore methods non-blocking
Wrap calls to the hardware key store provided by "platform_support" in a
"spawn_blocking()" method when called from the async context of the
Wallet. Since these calls access the OS, which in turn communicates with
the secure hardware, they may block. This is undesirable in an async
context and needs to be isolated in a separate thread.