Remove tokio JoinError from error enums
JoinError occurs either because of a panic or a cancellation. Since
spawn_blocking() does not support cancellation, we can assume that it
only occurs because of a panic. This means we can safely call
into_panic() on it, which itself panics when the JoinError does not
constitute a panic in the joined thread.