Process and Display unlock Result
* Introduce PinTimeoutScreenArgument Test

* Remove previously generated file

Now replaced with wallet_unlock_result.dart

* Remove async keyword

The sync_wallet_lock_status no longer has to be
async after passing in the wallet.

* Pass wallet to sync_wallet_lock_status

Pass on the wallet to make it more explicit that
the sync_wallet_lock_status() needs a lock on the
wallet.

* Introduce PinTimeoutScreenArgument

Align passing on the arguments by also using
an argument model for the PinTimeoutScreen.

Additional benefit is that it's now properly
serialized as a screen, meaning we can use
restorablePushReplacementNamed.

* Process PR Feedback

* Optimize formatter

* Singular naming for extensions

* Fix Rust warnings

* Use generic CheckPinResult

Map WalletUnlock results to a
CheckPinResult model on the flutter
side so that we can easily interpret
the results of both the unlock() and
confirmTransaction() calls.

* Rename PinVerify -> WalletUnlock (flutter)

Refactor the flutter code with the above rename,
the confirmTransaction() now also returns a
WalletUnlockXx model which is slightly undesired.
Need to look into if it's worth updating that at
this stage.

* Rename PinVerify -> WalletUnlock (rust)

Refactor the rust code with the above rename,
breaking the flutter app which will be fixed
in the next commit.

* Improve PinPage Tests

* Add PinPage Related Tests

* Improve PinBlockedScreen Tests

Also add a basic widget test, helps
with coverage.

* Improve PinTimeoutScreen Tests

Also add a basic widget test, helps
with coverage.

* Test PinTimeoutDescription

* Clean up Tests

Prefer instantiating WalletAppTestWidget
over the walletAppWrapper() function.

* Add PinOverlay Tests

Verify the overlay is shown when the wallet
state is locked and vice versa.

* Add IssuanceConfirmPinPage Tests

* Add ErrorScreen WidgetTest

Add explicit widget (non golden) test
to see if that is picked up by sonarqube
as coverage.

* Remove drop()

Remove the need to call drop(wallet) by
not assigning the wallet to a variable.

* Update filename

Filename didn't match the new classname

* Remove escape char in localization

* Clean up ErrorScreen

Combine .show() and showGeneric()

* Fix tests

* Unlock wallet by default

Set the is_locked flag to false after a successful
registration. This way any potential network
issue when calling the unlock api doesn't have
to be handled and since the user just created
a wallet (by providing the pin) this is also
not deemed a security risk, as it only unlocks
the local, empty, wallet.

* Update confirmTransaction()

Make it return PinVerifyResult so that
the confirm pin page can also render the
expected error states.

* Remove leftoverPinAttempts

Instead of having a dedicated method to request
the number of leftoverPinAttempts, this is now
always modeled in the return type of the
unlockWallet() method call. To simplify
the code I opted to remove it completely for now.
We might need to introduce some form of caching
for this value at some point, but I consider
that a future problem for now. Also updated
the mock with basic timeout/block result type support.

* Clean up Mocks

Clean up the mocks so that it's easier to use
Mocks.create() throughout the tests. The main
responsibility of the wallet_mocks class is
to provide the mocks (with little or no stubbing)
throughout the app. So that instantiating classes
under tests becomes trivial when we don't care what
a certain dependency does.

* Test leftOverPinAttempts

* Test PinBloc

* Add PinPage Screenshot Tests

* Add PinTimeoutScreen Screenshot Tests

* Add PinBlockedScreen Screenshot Tests

* Add ErrorScreen Screenshot Tests

* Expose isFinalAttempt

Expose a boolean flag so that the user can be
notified that the current pin attempt is really
the last attempt before being locked out of the
wallet completely.

* Localize (generic) ErrorScreen

* Localize PinBlockedScreen

* Localize PinTimeoutScreen

* Introduce PinBlockedScreen

* Add Basic PinTimeoutDescription

Introduce a the countdown widget to inform
the user when (s)he can retry entering the
pin.

* Introduce PinTimeoutScreen

* Introduce ErrorScreen

Introduce an error screen to which
we can navigate when pin verification
failed due to a generic error.

* Introduce PinVerifyResult

Instead of tying the result to the
wallet unlock call with the
WalletUnlockResult we now introduce
a PinVerifyResult so that future pin
related actions (like confirming a transaction)
can (hopefully) use the same model.

* Pass WalletUnlockResult to wallet_app

Pass on the WalletUnlockResult model to the
wallet app, so that the different error states
can eventually be rendered in the UI.

* Introduce WalletUnlock models

* Fix clippy warning

* Update wallet_app tests

* Change lock behavior

Locking should never fail, as this would
likely imply some broken security. As such
crash the app when it does.

* Run fmt & Add docs

* Add lock mechanism test

Verify that the lock()/unlock()
results in updated wallet.lock state

* Fix Deeplink Issue

Since the lock state is now coming out of the
wallet_core, there is also a slight delay since
it's coming over an async bridge. This delay is
just enough to trigger an issue in the deeplink
processing logic, causing the processUri method
to trigger before the device is unlocked (or actually
it triggers before it realizes it was ever locked).

To circumvent this, for now I added a fixed debounce
delay before checking if the app is locked. Would
be nice to avoid this delay in the future, but for
now opting this solution as sufficient since worst
case the deeplink is dropped or queued for later
processing. So no crucial (security) implications.

* Simply WALLET_LOCK_SINK reference

Instead of keeping it in a RwLock just
keep a direct reference to the StreamSink
as I currently don't oversee any concurrency
issues.

* Stream lock state from wallet_core

Make the init() method on the wallet_core
return a Stream that fires when the wallet
lock state changes.

* Implement placeholder (un)lock

* Define initial core (un)lock api