Provide config to AutoLockObserver
Turns out that we don't get a proper
callback when the wallet_core is
initialized (since we now return the
lock stream). As a temp solution now
delaying the config stream observation
with a 1 second delay. Plan is to restructure
the init call to make sure we don't have
to add this arbitrary delay.