Separate locked stream from init call
To know when the wallet is really initialized
(and avoid race conditions when trying to observe
e.g. the configuration struct stream) I opted change
the api of the init() call to no longer directly
return the locked stream. This way it completes
when the wallet is actually initialized. To still
be able to pass on locked events a new set/clear
combo has been added to the api, similar to the
configuration stream.