Fix secondary startup crash
* Re-add mocks

* Update formatting

* Update docs

* Fix startup crash

On Android the app would crash on startup
after it was previously closed with the back
button. This happened because the FlutterActivity
was killed (and thereby all knowledge about the
wallet_core being initialized was cleared) while
the actual Application (and by extend the native code)
was kept alive. This would cause the Flutter app
(on a subsequent startup) to try and initialize the
wallet_core again, which would now panic because it
was already initialized. To fix this, a new api
was introduced that allows the flutter code to check
whether or not the wallet_core is already initialized
and act based on that information.