Add animated pin loading state
* Show errors during pin setup

Make sure the error is also passed
into the PinField during the pin setup
flow, this makes sure the error animation
is also displayed here.

* Update PinField animation

Tweak the PinField animation to be more
similar to a wave instead of a snake.

Also tweak how the PinField behaves, namely:
- Never have the dots turn red.
- Clear the current input instantly when
an error occurts (don't keep the dots filled).

* Pre MR Cleanup

Clarify some code by assigning/renaming
variables.

* Add unlock mock delay

Add a short delay for wallet unlock in
the mock repositories so that the PinField
animations can be tested/seen on mock builds
of the app.

* Center StackedWalletCards

Noticed that during the onboarding
the cards weren't aligned properly,
since it's a minor change picking it
up on this branch.

* Make PinField state a required arg

And make sure to pass in the correct state
based on the BLoCs current state.

* Move PinFieldDemo to design system screen

* Clean up PinField widget

* Add error shake animation

* Add Pinfield loading animation

* Setup temp PinField playground