Hide content in app switcher
* Update PrivacyCover

* Fix import

* Hide content in app switcher (iOS)

Draw the app logo on top of whatever
app content was visible to hide the
content for privacy reasons.

Also slightly reorganized the root
dependency widgets to keep the main.dart
a bit more clean.

* Hide content in app switcher (Android)

Hide the contents of the wallet app in
the preview shown in the app switcher
for privacy reasons. This is done by
setting the FLAG_SECURE when the app
enters the background. The flag is cleared
in the onResume to allow for screenshots
when the app is in the foreground.