Upgrade to Dart 3.0
* Add proper _buildError to CardOverviewScreen

* Fix formatting

* Remove unused import

* Migrate CardOverviewState to sealed

* Migrate CardHistoryState to sealed

Also add a basic error state as that
was previously unhandled.

* Migrate CardDataState to sealed

Also add a simple error state, as that
was previously unhandled.

* Migrate IssuanceState to sealed

* Migrate HistoryOverviewState to sealed

Also added a basic implementation for HistoryOverviewLoadFailure,
as it was not yet covered.

* Migrate ChangeLanguageState to sealed

* Migrate HomeState to sealed

Also simplified the HomeBloc a bit
by removing the internal state in favor
of a uid.

The HomeScreen was also simplified as the
forced refresh is no longer needed here
since the menu is now nested (i.e. no
need to navigate up when pressing the
tab again).

* Migrate QrScanState to sealed

* Migrate PinState to sealed

* Migrate OrganizationDetailState to sealed

* Migrate MockDigidState to sealed

* Migrate HistoryDetailState to sealed

Also noticed that the failure state was
not implemented / supported so added a
generic error page.

* Migrate CardSummaryState to sealed

* Migrate WalletPersonalizeState to sealed

* Migrate VerificationState to sealed

* Migrate SplashState to sealed

* Migrate SignState to sealed

* Migrate SetupSecurityState to sealed

* Migrate FlashlightState to sealed

* Migrate MenuState to sealed

* Update to Dart 3.0

Also update dependencies so that the
generation still works with the dart
upgrade.