Update Goldens
* Update VerificationConfirmPinPage Goldens

Update the goldens to reflect the more
responsive pin entry layout.

* Update SplashScreen Goldens

Update the goldens to reflect the new
primary color used in the wallet logo.

* Update IssuanceConfirmPinPage Goldens

Update the goldens to reflect the new
more responsive pin layout.

* Update SetupSecurity Goldens

Update the goldens to reflect the new
primary color used in the wallet logo.

* Update PinTimeoutScreen Goldens

Update the goldens to reflect the new
action button theme color.

* Update WalletPersonalizePage Goldens

To make the screen more responsive the
layout was slightly tweaked, pushing the
PinField down a bit. This is now captured by
the screenshot master image as well.

* Update PinPage Goldens

To make the screen more responsive the
layout was slightly tweaked, pushing the
PinField down a bit. This is now captured by
the screenshot master image as well.

* Update SignConfirmPinPage Goldens

To make the screen more responsive the
layout was slighlty tweaked, pushing the
PinField down a bit. This is now captured by
the screenshot master image as well.

* Update ErrorScreen Golden

The close button's color was updated,
updating the master image accordingly.