Tweak lokalisation setup
* Add script subhead

* Update lokalise.md

* Fix imports

* Use AppLocalizations.localizationsDelegates

Revert back to using the AppLocalizations.localizationsDelegates
as this is provided (was removed when using the lokalise plugin).

* Update lokalise docs

Update the docs to indicate that modified keys will be replaced
with these default settings, which can be disabled by setting
the flag to false.

* Update upload script

Make it more scalable by looping
over the locales.

* Remove formatting rule

Remove the formatting since the translations
live in a synthetic package and thus is not
part of our project / formatting.

* Remove arb_utils dependency

Since lokalise exports are sorted when they are exported,
the arb_utils dependency and the script that relies on it
(arb_sort) can be removed in favor of re-importing from
lokalise once in a while.

* Update lokalise.md

Add scripts to easily sync the translations
with lokalise.

* Revert to l10n synthetic-package

* Remove flutter_intl

Remove the plugin because it messes with the formatting
on every build command.

* Updaate lokalise docs

* Move localizations & enable flutter_intl

Move the localizations from lib/src/localization
to the more common lib/l10n directory.

Also enabled the flutter_intl package available
for VsCode and IntelliJ IDEs to make it trivial
to extract new translations into the ARB files.
See: https://plugins.jetbrains.com/plugin/13666-flutter-intl
Usage of this plugin is optional.