Update onboarding ui
* Fix formatting issues

* Remove lock icon from dashboard

Opted to keep the CardOverviewLockWalletPressed
event for now locking from the CardOverview might
still be desired in some way/shape or form.

* Add //FIXME note

* Update InfoRow

Update the InfoRow to make it more
flexible when it comes to text styling,
it now takes Text widgets instead of
Strings directly (similar to other framework
widgets) so that the style can be overriden by
the creator. This is used in e.g. the
WalletPersonalizeDataIncorrectScreen.

* Update WalletPersonalizeDataIncorrectScreen

Clean up copy and add municipality section.

* Update GenericLoadingPage

- Make it re-usable for the WalletPersonalizeScreen
loading states
- Make it so the loading indicator is centered in
the screen (as per the designs)
- Add an optional cancel button
- Update the tests to verify the new UI

* Update setupSecurityCompletedPageDescription

* Update SetupSecurityCreatingWallet ui

Update the loader shown when the
SetupSecurityScreen is rendering the
SetupSecurityCreatingWallet state.

* Introduce GenericLoadingPage