Improve wallet_app /feature/common/* tests
* Revert pbxproj changes

* Introduce VersionText Test

* Introduce StackedWalletCards Test

Spend quite a while trying to optimize the
StackedWalletCards so that the cards don't
shrink down when there is not enough vertical
space. But positioning the widgets directly
in the Stack causes the Stack to be unable
to size itself. And calculating the size
upfront is tricky because it depends on the
horizontal space. As such sticking to the
current solution since it works and is
straight forward.

* Introduce StackedWalletCards Test

* Introduce SliverDivider Test

* Introduce SelectCardRow Test

* Fix formatting

* Remove unused goldens

* Fix InfoRow analyse warning

* Introduce SvgOrImage Test

* Introduce InfoRow Test

* Introduce DocumentSection test

* Introduce MaxBrightness test

* Introduce LimitFontScaling test

* Introduce DoOnInit test

* Fix formatting

* Introduce DataAttributeSection Tests

* Introduce ExplanationSheet tests

* Introduce ConfirmActionSheet tests

* Add landscape placeholder screen test

* Introduce PlaceholderScreen Test

* Introduce FlowTerminalPage Test

* Ignore generated coverage report

Generated with:

flutter test --coverage
genhtml coverage/lcov.info -o coverage/html

* Introduce CheckDataOfferingPage Test

* Fix IntroductionScreen Test

Tweak some paddings to match the Figma design

* Move Page/Screen/Sheet

Move widgets that represent a page, screen
or sheet to a dedicated directory.

* Fix test formatting

* Add CenteredLoadingIndicator Test

* Add AnimatedLinearProgressIndicator Test

* Add PolicySection Test

* Remove ExtendedPolicyRow

As it turned out to be unused.

* Add PolicyRow Test

* Add OrganizationLogo Test

* Add TimelineSectionSliver Test

* Add TimelineSectionHeader Test

* Add TimelineAttributeRow Test

* Update WalletCardItem docs

* Flutter analyze fix

* Move WalletCardItem Tests

To match the package in the main project

* Add test template to README.md

* Add AttributeRow Test

* Add TextIconButton Test

* Add LinkButton Test

* Add ConfirmButtons Test

* Fix AnimatedVisibilityBackButton test

* Add AnimatedVisibilityBackButton test

* Re-import translations

* Test BottomBackButton