Fix introduction navigation issue
* Fix introduction navigation issue (2)

Improve the builder a logic a little bit by relying
on the cached value, this also resolves a runtime warning
when the user uses the skip button.

* Fix introduction navigation issue

Fix the issue where navigation from the introduction
screen fails (when not in landscape) because the
Overlay is trying to reference an unmounted widget
to position itself.