Fix locking mechanism
* Fix locking mechanism

When the app is put in the background and killed by the
host OS the AutoLockObserver will not lock the wallet
according to the locking mechanism. In this case we now
fall back to simply locking the wallet to make sure the
user can not use this a workaround to keep the wallet
unlocked.