Android privacy & mock build fix
* Run formatter

* Re-introduce onResume/onPause

Also set the flags in onResume and onPause
for proper support of android versions <Q

* Fix mock build

While walking through the app noticed that the mock
digid flow was broken because it now relies on the
DigidAuthRepository which wasn't properly mocked. The
same goes for the TypedWalletCore (not mocked). This
commit introduces mocks for those classes that suffice
the minimum requirements to get the mock build of the
app to function as expected again.

* Android PrivacyCover fix

Privacy cover did not always (i.e. on all platform versions)
show up  correctly when setting it in the onPause callback, now
relying on the onWindowFocusChanged instead.