Update error screens
* Update copy & goldens

* Perform minor cleanup

- Clean up docs
- Add proper todos
- Make statusCode nullable to explicitly communicate
it might not exists.

* Update english generic error copy

* Fix analyze & formatting issues

* Introduce HelpSheet tests

* Update copy & related tests

* Update PinPage test

Re-create golden with updated error
screen and introduce golden for the
generic error page.

* Fix CardDataScreen test

Re-create golden with updated copy

* Fix CardOverviewScreen test

Re-create golden with updated copy

* Fix CardHistoryScreen test

Re-create golden with updated copy

* Fix IntroductionScreen test

Re-create golden with updated image

* Fix HistoryDetailScreen test

Re-create golden with new (updated) copy

* Fix HistoryOverviewScreen test

Update copy changes are now reflected
correctly.

* Fix ThemeScreen tests

* Improve ErrorScreen test

Test the helper methods to render
the error screen for a specific error,
namely showGeneric and showServer.

* Add mock unlock paths

On debug builds of the app we now map the pins
111111,222222,333333,444444,555555 to the different
error paths that can occur when unlocking the app
to allow us to easily demo and verify the UI related
to these errors.

* Add error illustrations

Add the illustrations for the general and
server error screens.

* Introduce HelpSheet and translations

* Split out CheckPin server & generic errors