Remove "hardware" feature of platform_support crate
Hardware keys are now always enabled and exported by this crate, since otherwise there would be no need for the user of this crate to use it at all.
The wallet crate now references the hardware keys by name in init.rs, and elsewhere it is generic over the traits that these types implement.