Update onboarding and introduce expectations screen
* Fix typo

* Pin PageView scroll physics

To avoid getting a lot of whitespace when trying
to scroll the page view to page ''-1' or 'nrPages + 1'
we now force the clamping scrollPhysics which disallows
any scroll in the overscroll direction.

* Fix pageIndicator accessibility rectangle

The highlight rectangle was not rendered correctly anymore
due to a change in the amount of pages.

* Fix formatting

* Fix golden tests

The onboarding has been shortened by one page
and a new button was added to the theme screen.

* Fix ReportIssueScreen issue

Noticed the show method was not using the
correct route, which could allow displaying
this route while the wallet is locked (which
is undesired for this screen). Fixed by wrapping
it in a SecuredPageRoute.

* Clean up IntroductionScreen

* Add missing lokalise key

* Introduce IntroductionExpectationsScreen golden test

* Lokalise IntroductionExpectationsScreen

* Update IntroductionExpectationsScreen ui

* Fix scrollController not attached

When using the next button to navigate to the
next screen of the onboarding the calculation
for the pager did not completely correctly
because the scroll controller wasn't attached
with clients yet. This is fixed by checking this
before retrieving the current position.

* Extract RoundedBackButton

* Handle (impossible?) scroll edge case

Make sure that if the cache is filled while
the content is scrolled (unlikely, maybe even
impossible) that this is taken into account so
that we never cache an invalid value.

* Simplify IntroductionScreen

Build the paging indicator inside of the stack.

* Introduce initial IntroductionExpectationsScreen

* Improve PagingIndicator animations

Improve the behaviour of the PagingIndicator
while scrolling vertically on smaller screens.
It now moves together with the text and fades
out as the scroll distance increases.

* Make onboarding image sticky