DigiD Termination & Help flow (ui-only)
* Tweak variable naming

* Fix cancel dialog race condition

When loggin into Digid and pressing the cancel button *just* before the
state changes into something else, the click is processed after this
state change. This means the dialog is shown on top of the next instead
of the current state, which in term means our attempt to close dialogs
on a state change is now in vein, since the dialog doesn't exist yet
at that point.

To fix this issue we now check for the current state inside the click
handler that spawns the dialog. If the state is not in any of the
expected states where we would possibly have to show this dialog, the
call is ignored.

* Verify WalletPersonalizeConnectDigid state

Verify the WalletPersonalizeConnectDigid state's ui and check if check
if the digid url is correctly opened.

* Verify dialog dismiss behaviour

Add a test which verifies that the stop digid login dialog is correctly
dismissed when a result comes in as it's being shown.

* Improve dialog dismiss logic

While considering writing a test to verify this behaviour I realized
that I assume the screen is opened as a named route, which it might not
be. The improved logic only checks for dialogs to avoid this assumption.

* Fix DigiD dialog bug

This fixes a bug where:
1.The user attempts to login with digid
2 The user is redirected back into the app
3. Processing the DigiD redirect url takes some time so the user
clicks cancel and the cancel login dialogs shows up
4. While the dialog is shown the digid login result comes in and the
app navigates to show the result
5. The app is now in a success state, but the cancel dialog is shown
on top of this state

This fix closes any visible dialog at step 4 to circumvent this issue

* Fix typo in introductionPage2Title

* Revert translation updates

This branch still relies on other translations that were accidentally
replaced in a pull from lokalise. Reverted with this commit.

* Update formatting

* Fix authenticating state typo

* Add DigiD failure illustration

* Add DigidHelpScreen golden test

* Launch externally

Make sure the uris are opened in the
external browser.

* Launch digid uris

When clicking the request help or request
digid buttons forward the user to the correct
page in the platform browser.

* Update DigidHelpScreen lokalisations

* Setup initial DigidHelpScreen

* Fix WalletPersonalizeScreen navigation

Temporarily tweaked the navigation to more
easily test the dialog flow, reverting with this
commit.

* Add StopDigiDDialog golden

Verify the dialog shows up as expected (and
when pressing the cancel button).

* Update (sheet related) goldens

The tintSurface changed, causing the golden tests to
fail as expected. These are updated to reflect the new
truth with this commit.