Fix navigation bug
* Fix navigation bug

After the initial setup the 'back' button is shown on the home screen
and the user can navigate back because the introduction screen wasn't
properly dismissed. Now making sure we pop upto the splashRoute when
navigating away from the onboarding, so that the user also can't return
using a physical back button during any part of the onboarding.